Refactored and commented a lot of code
Also added the ability to restart the game using the 'r' key.

There is currently a memory leak with the physics box and physics circle class. It doesn't delete the mesh and shader it creates in its constructor. I'm not sure why, but when we call delete on the mesh and shader it creates, it fails. I think it's because they belong to the base class.